todos의 남은 개수, 종료 개수를 담은 객체로 수정하라

+++ app-server/src/main/java/com/growth/task/task/dto/TaskTodoResponse.java
@@ -5,14 +5,17 @@ import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+/**
+ * Task의 Todos 진행률 DTO
+ */
 @NoArgsConstructor
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 @Getter
 public class TaskTodoResponse {
-    private Long remain;
-    private Long done;
+    private int remain;
+    private int done;
 
-    public TaskTodoResponse(Long remain, Long done) {
+    public TaskTodoResponse(int remain, int done) {
         this.remain = remain;
         this.done = done;
     }
